Added information as per #117

+++ Aikuma/src/org/lp20/aikuma/audio/record/Recorder.java
@@ -90,7 +90,7 @@ public class Recorder implements AudioHandler, MicrophoneListener, Sampler {
 		return file.getCurrentSample();
 	}
 
-	public float getCurrentMsec() {
+	public int getCurrentMsec() {
 		return sampleToMsec(getCurrentSample());
 	}
